requirements.txt
# requirements.txt
# Required packages for Invoice Processing Application

# Google API packages
gspread>=5.7.0
google-auth>=2.16.0
google-auth-oauthlib>=0.8.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.70.0

# Data processing
pandas>=1.5.0
numpy>=1.21.0

# Image processing
Pillow>=9.0.0
requests>=2.28.0

# PDF generation
reportlab>=3.6.0

# GUI (usually comes with Python)
# tkinter - included with Python standard library